hosted-old/requirements.txt
# Content Scout — Hosted Agent Dependencies

# Foundry hosting adapter (Responses + Invocations protocols).
# NOTE: Only pre-release builds exist today — install with `pip install --pre`
# against the latest Agent Framework release.
agent-framework>=1.0.0a1
agent-framework-foundry-hosting>=1.0.0a1

# Azure identity — Managed Identity / DefaultAzureCredential (zero secrets)
azure-identity>=1.17.0

# Azure Key Vault — scanner API keys loaded at startup
azure-keyvault-secrets>=4.8.0

# Azure Blob Storage — persistent reports + dedup tracker
azure-storage-blob>=12.22.0

# Azure OpenAI / Foundry model calls via MI token provider
openai>=1.54.0

# HTTP client used by scanners
httpx>=0.27.0

# Environment loader (local dev only)
python-dotenv>=1.0.0
